requirements.txt
django-crispy-forms
django
asgiref==3.2.3
beautifulsoup4==4.8.2
boto==2.49.0
boto3==1.11.12
botocore==1.14.12
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
django-debug-toolbar==2.2
djangorestframework==3.11.0
docutils==0.15.2
idna==2.8
jmespath==0.9.4
Pillow==7.0.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
s3transfer==0.3.3
six==1.14.0
soupsieve==1.9.5
sqlparse==0.3.0
urllib3==1.25.8
